Boat rework
We finally decided on how to move the boat
For first build, we will just have the boat move forward.  Waves will be controlled by shader
Eventually, we will rework movement (again) to follow the spline, but this will do for first build
