+++ Assets/General/Art/waterboatmaterialtest.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a8cb6cef943b0064a85609e52e906903
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
