+++ Assets/General/Scripts/BoatMovement/BoatMover.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6704926a89f546e40b13b69345c4f895
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
