+++ Assets/General/Scripts/BoatMovement/BoatPassengerTransfer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7b98f385d4a4e8f47b41c96bf7d57493
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
