+++ Assets/General/Scripts/BoatMovement/BoxDroppingScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4a86df4a4da9063479d67aca12d24105
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
